fix weekly notes len
